# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: listings
    description: Public business listings
  - name: bookings
    description: Booking management
  - name: customer-auth
    description: Customer authentication
  - name: customer
    description: Customer profile and history
  - name: ratings
    description: Post-service ratings
  - name: waitlist
    description: Waitlist management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Browse businesses
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: lat
          schema: { type: number }
        - in: query
          name: lng
          schema: { type: number }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"

  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get business detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Business detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "404":
          description: Not found

  /listings/{id}/slots:
    get:
      operationId: getListingSlots
      tags: [listings]
      summary: Get available slots for a service on a date
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: serviceId
          required: true
          schema: { type: string }
        - in: query
          name: date
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlotsResponse"

  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (guest or authenticated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingCreatedResponse"

  /bookings/by-token/{token}:
    get:
      operationId: getBookingByToken
      tags: [bookings]
      summary: Look up a booking by its public token
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
        "404":
          description: Not found

  /auth/customer/otp:
    post:
      operationId: sendCustomerOtp
      tags: [customer-auth]
      summary: Send OTP to customer phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string }
              required: [phone]
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  testOtp: { type: string }

  /auth/customer/verify:
    post:
      operationId: verifyCustomerOtp
      tags: [customer-auth]
      summary: Verify OTP and return JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string }
                otp: { type: string }
              required: [phone, otp]
      responses:
        "200":
          description: JWT token and user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /customer/bookings:
    get:
      operationId: getCustomerBookings
      tags: [customer]
      summary: Get authenticated customer booking history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookingDetail"

  /ratings:
    post:
      operationId: submitRating
      tags: [ratings]
      summary: Submit a post-service rating using one-time token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
              required: [token, rating]
      responses:
        "200":
          description: Rating submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }

  /waitlist:
    post:
      operationId: joinWaitlist
      tags: [waitlist]
      summary: Join a waitlist for a fully booked slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistBody"
      responses:
        "201":
          description: Added to waitlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  position: { type: integer }
                  waitlistId: { type: string }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ServiceSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        durationMin: { type: integer }
        price: { type: integer }
        description: { type: string, nullable: true }
      required: [id, name, durationMin, price]

    ListingSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        image: { type: string }
        averageRating: { type: number }
        totalRatings: { type: integer }
        distanceKm: { type: number, nullable: true }
        city: { type: string }
        address: { type: string }
        openNow: { type: boolean }
        waitTimeMinutes: { type: integer }
        queueCount: { type: integer }
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceSummary"
      required: [id, name, category, image, averageRating, totalRatings, openNow, services]

    ListingsResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/ListingSummary"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
      required: [listings, total, page, limit]

    ReviewSummary:
      type: object
      properties:
        id: { type: string }
        rating: { type: integer }
        comment: { type: string }
        customerName: { type: string }
        date: { type: string }

    ListingDetail:
      allOf:
        - $ref: "#/components/schemas/ListingSummary"
        - type: object
          properties:
            description: { type: string }
            phone: { type: string, nullable: true }
            website: { type: string, nullable: true }
            hours: { type: string }
            hoursDetail: { type: object, nullable: true }
            amenities:
              type: array
              items: { type: string }
            staff:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  name: { type: string }
                  role: { type: string }
                  rating: { type: number }
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/ReviewSummary"

    SlotItem:
      type: object
      properties:
        time: { type: string }
        available: { type: boolean }
        spotsLeft: { type: integer }
      required: [time, available, spotsLeft]

    SlotsResponse:
      type: object
      properties:
        slots:
          type: array
          items:
            $ref: "#/components/schemas/SlotItem"
        date: { type: string }
        serviceId: { type: string }
        durationMin: { type: integer }
      required: [slots, date, serviceId, durationMin]

    CreateBookingBody:
      type: object
      properties:
        listingId: { type: string }
        serviceId: { type: string }
        staffId: { type: string }
        resourceId: { type: string }
        date: { type: string }
        time: { type: string }
        persons: { type: integer, default: 1 }
        seats:
          type: array
          items: { type: string }
        notes: { type: string }
        isQueueJoin: { type: boolean }
        customerName: { type: string }
        customerPhone: { type: string }
        paymentId: { type: string }
      required: [listingId, date, time]

    BookingDetail:
      type: object
      properties:
        id: { type: string }
        token: { type: string }
        ratingToken: { type: string, nullable: true }
        status: { type: string }
        date: { type: string }
        time: { type: string }
        persons: { type: integer }
        platformFee: { type: integer }
        businessName: { type: string }
        businessAddress: { type: string }
        businessImageUrl: { type: string }
        serviceName: { type: string }
        createdAt: { type: string }

    BookingCreatedResponse:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/BookingDetail"
        token: { type: string }
        platformFee: { type: integer }
        serviceFeeAmount: { type: integer }
        status: { type: string }
        paymentId: { type: string, nullable: true }

    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user:
          type: object
          properties:
            id: { type: string }
            name: { type: string, nullable: true }
            phone: { type: string }
            email: { type: string, nullable: true }
      required: [token, user]

    WaitlistBody:
      type: object
      properties:
        listingId: { type: string }
        serviceId: { type: string }
        date: { type: string }
        time: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
      required: [listingId, date, time, customerName, customerPhone]